GCW-3314 updated designation overlay
