+++ app/components/goodcity/designate-link.js
@@ -1,6 +1,5 @@
 import Ember from "ember";
 import _ from "lodash";
-import DesignationActions from "stock/mixins/designation_actions";
 
 //
 // Small link component which holds all the designation actions
@@ -11,4 +10,6 @@ import DesignationActions from "stock/mixins/designation_actions";
 //    <div> Click me to designate {{pkg.inventory_number}} to {{orderToDesignateTo.code}}
 // {{/goodcity/designate-link}}
 //
-export default Ember.Component.extend(DesignationActions, {});
+export default Ember.Component.extend({
+  designationService: Ember.inject.service()
+});
