+++ app/components/goodcity/order-block.js
@@ -1,0 +1,3 @@
+import Ember from "ember";
+
+export default Ember.Component.extend({});
