+++ app/styles/_search.scss
@@ -223,7 +223,8 @@
 
 // User Search Page
 .organisations_search,
-.users-search {
+.users-search,
+.orders-search {
   .back-link {
     margin-top: 0.5rem;
   }
