+++ app/templates/_global_components.hbs
@@ -1,11 +1,24 @@
 {{!-- Service-backed components --}}
 
 {{#let-alias designationService as |s|}}
+  {{#if s.displayDesignateForm}}
+    {{goodcity/designate-form
+        quantity=s.designationQty
+        maxQuantity=s.designatableQuantity
+        confirmAction=s.completeDesignation
+        cancelAction=s.cancelDesignation
+        order=s.designationTargetOrder
+        pkg=s.designationTargetPackage
+        readyToDesignate=s.readyToDesignate
+        showItemAvailability=s.showItemAvailability
+      }}
+  {{/if}}
+
   {{goodcity/orders-search-overlay
-    open=s.openOrderSearch
-    onSelect=s.onOrderSelected
-    searchProps=s.orderSearchProps
-  }}
+      open=s.openOrderSearch
+      onSelect=s.onOrderSelected
+      searchProps=s.orderSearchProps
+    }}
 {{/let-alias}}
 
 {{#let-alias organisationService as |s|}}
