+++ app/templates/components/goodcity/designate-link.hbs
@@ -1,17 +1,5 @@
 <div class="designate-link">
-  <div {{action 'beginDesignation' pkg order}}>
+  <div {{action 'beginDesignation' pkg order target=designationService}}>
     {{yield}}
   </div>
-
-  {{#if readyToDesignate}}
-    {{goodcity/designate-form
-      quantity=designationQty
-      maxQuantity=designatableQuantity
-      editableQty=editableQty
-      confirmAction=(action "completeDesignation")
-      cancelAction=(action "cancelDesignation")
-      order=designationTargetOrder
-      pkg=designationTargetPackage
-    }}
-  {{/if}}
 </div>
